+++ internal/ray/ray.go
@@ -1,6 +1,6 @@
 package ray
 
-import "github.com/flynn-nrg/izpi/internal/vec3"
+import "github.com/flynn-nrg/go-vfx/math32/vec3"
 
 // Ensure interface compliance.
 var _ Ray = (*RayImpl)(nil)
@@ -9,12 +9,12 @@ var _ Ray = (*RayImpl)(nil)
 type RayImpl struct {
 	origin    *vec3.Vec3Impl
 	direction *vec3.Vec3Impl
-	lambda    float64
-	time      float64
+	lambda    float32
+	time      float32
 }
 
 // New returns a new ray with the supplied origin and direction vectors and time.
-func New(origin *vec3.Vec3Impl, direction *vec3.Vec3Impl, time float64) *RayImpl {
+func New(origin *vec3.Vec3Impl, direction *vec3.Vec3Impl, time float32) *RayImpl {
 	return &RayImpl{
 		origin:    origin,
 		direction: direction,
@@ -23,7 +23,7 @@ func New(origin *vec3.Vec3Impl, direction *vec3.Vec3Impl, time float64) *RayImpl
 }
 
 // NewWithLambda returns a new ray with the supplied origin and direction vectors, time, and wavelength.
-func NewWithLambda(origin *vec3.Vec3Impl, direction *vec3.Vec3Impl, time float64, lambda float64) *RayImpl {
+func NewWithLambda(origin *vec3.Vec3Impl, direction *vec3.Vec3Impl, time float32, lambda float32) *RayImpl {
 	return &RayImpl{
 		origin:    origin,
 		direction: direction,
@@ -43,21 +43,21 @@ func (r *RayImpl) Direction() *vec3.Vec3Impl {
 }
 
 // PointAtParameter is used to traverse the ray.
-func (r *RayImpl) PointAtParameter(t float64) *vec3.Vec3Impl {
+func (r *RayImpl) PointAtParameter(t float32) *vec3.Vec3Impl {
 	return vec3.Add(r.origin, vec3.ScalarMul(r.direction, t))
 }
 
 // Time returns the time associated with this ray.
-func (r *RayImpl) Time() float64 {
+func (r *RayImpl) Time() float32 {
 	return r.time
 }
 
 // Lambda returns the wavelength associated with this ray.
-func (r *RayImpl) Lambda() float64 {
+func (r *RayImpl) Lambda() float32 {
 	return r.lambda
 }
 
 // SetLambda sets the wavelength associated with this ray.
-func (r *RayImpl) SetLambda(lambda float64) {
+func (r *RayImpl) SetLambda(lambda float32) {
 	r.lambda = lambda
 }
